Fix some issues due to merge
