+++ source/Zodiac_compatibility/data/dataArray.js
@@ -1,152 +1,32 @@
-const zodiacDateRangesLeft = [
-  [0, "Dec 22 - Jan 19"], // Capricorn
-  [360, "Dec 22 - Jan 19"], // Capricorn
-
-  [30, "Nov 22 - Dec 21"], // Sagittarius
-  [-330, "Nov 22 - Dec 21"], // Sagittarius
-
-  [60, "Oct 23 - Nov 21"], // Scorpio
-  [-300, "Oct 23 - Nov 21"], // Scorpio
-
-  [90, "Sep 23 - Oct 22"], // Libra
-  [-270, "Sep 23 - Oct 22"], // Libra
-
-  [120, "Aug 23 - Sep 22"], // Virgo
-  [-240, "Aug 23 - Sep 22"], // Virgo
-
-  [150, "Jul 23 - Aug 22"], // Leo
-  [-210, "Jul 23 - Aug 22"], // Leo
-
-  [180, "Jun 21 - Jul 22"], // Cancer
-  [-180, "Jun 21 - Jul 22"], // Cancer
-
-  [210, "May 21 - Jun 20"], // Gemini
-  [-150, "May 21 - Jun 20"], // Gemini
-
-  [240, "Apr 20 - May 20"], // Taurus
-  [-120, "Apr 20 - May 20"], // Taurus
-
-  [270, "Mar 21 - Apr 19"], // Aries
-  [-90, "Mar 21 - Apr 19"], // Aries
-
-  [300, "Feb 19 - Mar 20"], // Pisces
-  [-60, "Feb 19 - Mar 20"], // Pisces
-
-  [330, "Jan 20 - Feb 18"], // Aquarius
-  [-30, "Jan 20 - Feb 18"], // Aquarius
+const zodiacOrder = [
+  "Capricorn",
+  "Sagittarius",
+  "Scorpio",
+  "Libra",
+  "Virgo",
+  "Leo",
+  "Cancer",
+  "Gemini",
+  "Taurus",
+  "Aries",
+  "Pisces",
+  "Aquarius",
 ];
-const zodiacDateRangesRight = [
-  [0, "Jun 21 - Jul 22"], // Cancer
-  [360, "Jun 21 - Jul 22"], // Cancer
-
-  [30, "May 21 - Jun 20"], // Gemini
-  [-330, "May 21 - Jun 20"], // Gemini
-
-  [60, "Apr 20 - May 20"], // Taurus
-  [-300, "Apr 20 - May 20"], // Taurus
-
-  [90, "Mar 21 - Apr 19"], // Aries
-  [-270, "Mar 21 - Apr 19"], // Aries
-
-  [120, "Feb 19 - Mar 20"], // Pisces
-  [-240, "Feb 19 - Mar 20"], // Pisces
-
-  [150, "Jan 20 - Feb 18"], // Aquarius
-  [-210, "Jan 20 - Feb 18"], // Aquarius
-
-  [180, "Dec 22 - Jan 19"], // Capricorn
-  [-180, "Dec 22 - Jan 19"], // Capricorn
-
-  [210, "Nov 22 - Dec 21"], // Sagittarius
-  [-150, "Nov 22 - Dec 21"], // Sagittarius
-
-  [240, "Oct 23 - Nov 21"], // Scorpio
-  [-120, "Oct 23 - Nov 21"], // Scorpio
-
-  [270, "Sep 23 - Oct 22"], // Libra
-  [-90, "Sep 23 - Oct 22"], // Libra
-
-  [300, "Aug 23 - Sep 22"], // Virgo
-  [-60, "Aug 23 - Sep 22"], // Virgo
-
-  [330, "Jul 23 - Aug 22"], // Leo
-  [-30, "Jul 23 - Aug 22"], // Leo
-];
-const mappingLeft = [
-  [0, "Capricorn"],
-  [360, "Capricorn"],
-
-  [30, "Sagittarius"],
-  [-330, "Sagittarius"],
-
-  [60, "Scorpio"],
-  [-300, "Scorpio"],
-
-  [90, "Libra"],
-  [-270, "Libra"],
-
-  [120, "Virgo"],
-  [-240, "Virgo"],
-
-  [150, "Leo"],
-  [-210, "Leo"],
-
-  [180, "Cancer"],
-  [-180, "Cancer"],
-
-  [210, "Gemini"],
-  [-150, "Gemini"],
-
-  [240, "Taurus"],
-  [-120, "Taurus"],
-
-  [270, "Aries"],
-  [-90, "Aries"],
-
-  [300, "Pisces"],
-  [-60, "Pisces"],
-
-  [330, "Aquarius"],
-  [-30, "Aquarius"],
-];
-const mappingRight = [
-  [0, "Cancer"],
-  [360, "Cancer"],
-
-  [30, "Gemini"],
-  [-330, "Gemini"],
-
-  [60, "Taurus"],
-  [-300, "Taurus"],
-
-  [90, "Aries"],
-  [-270, "Aries"],
-
-  [120, "Pisces"],
-  [-240, "Pisces"],
-
-  [150, "Aquarius"],
-  [-210, "Aquarius"],
-
-  [180, "Capricorn"],
-  [-180, "Capricorn"],
-
-  [210, "Sagittarius"],
-  [-150, "Sagittarius"],
-
-  [240, "Scorpio"],
-  [-120, "Scorpio"],
-
-  [270, "Libra"],
-  [-90, "Libra"],
-
-  [300, "Virgo"],
-  [-60, "Virgo"],
-
-  [330, "Leo"],
-  [-30, "Leo"],
-];
-const romantic = [
+const zodiacDateRanges = {
+  Capricorn: "Dec 22 - Jan 19",
+  Sagittarius: "Nov 22 - Dec 21",
+  Scorpio: "Oct 23 - Nov 21",
+  Libra: "Sep 23 - Oct 22",
+  Virgo: "Aug 23 - Sep 22",
+  Leo: "Jul 23 - Aug 22",
+  Cancer: "Jun 21 - Jul 22",
+  Gemini: "May 21 - Jun 20",
+  Taurus: "Apr 20 - May 20",
+  Aries: "Mar 21 - Apr 19",
+  Pisces: "Feb 19 - Mar 20",
+  Aquarius: "Jan 20 - Feb 18",
+};
+const romantic = new Map([
   [
     "Aries and Taurus",
     "Aries and Taurus can experience an intense and passionate romantic relationship. Aries brings excitement, spontaneity, and a sense of adventure, while Taurus offers stability, loyalty, and sensuality. However, conflicts may arise due to Aries' impulsive nature and Taurus' desire for security and routine. Open communication and compromise are key to maintaining a harmonious and fulfilling romance.",
@@ -459,11 +339,5 @@ const romantic = [
     "Pisces and Pisces",
     "When two Pisces individuals come together in a romantic relationship, they create a deep and emotionally rich connection. Both partners have a sensitive and compassionate nature, making them highly attuned to each other's emotional needs. They understand and empathize with each other's emotions, creating a safe and nurturing space for love to flourish. Pisces partners share a strong spiritual and intuitive connection, often feeling like they can understand each other on a profound level without the need for words. They enjoy creating a romantic and dreamy atmosphere, filled with love, romance, and imaginative gestures. However, their deep emotional sensitivity can sometimes lead to emotional intensity or mood swings within the relationship. It's important for Pisces partners to practice effective communication, establish healthy boundaries, and support each other's emotional well-being. When they embrace their shared empathy, emotional depth, and spiritual connection, Pisces partners can create a truly magical and soulful romantic bond.",
   ],
-];
-export {
-  zodiacDateRangesLeft,
-  zodiacDateRangesRight,
-  mappingLeft,
-  mappingRight,
-  romantic,
-};
+]);
+export { zodiacOrder, zodiacDateRanges, romantic };
