+++ source/Zodiac_compatibility/index.html
@@ -26,7 +26,7 @@
   <body>
     <h1 class="title">Love Compatibility</h1>
     <div id="guide">
-      <button id="how_to">How to Use</button>
+      <button class="button" id="how_to">How to Use</button>
     </div>
     <div class="container">
       <div class="wheel_field" id="left_wheel_field">
